+++ src/main/java/pro/sky/java/course2/calculator/service/CalcService.java
@@ -4,5 +4,5 @@ public interface CalcService {
     int getSum(int num1, int num2);
     int getDifference(int num1, int num2);
     int getComposition(int num1, int num2);
-    float getDivision(int num1, int num2);
+    float getDivision(float num1, float num2);
 }
